% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_github.R
\name{build_github}
\alias{build_github}
\alias{download_github_zip}
\title{Build source package from source code in a GitHub repository.}
\usage{
build_github(urls, destdir = ".")

download_github_zip(urls, destdir)
}
\arguments{
\item{urls}{Character vector giving GitHub repositories with the
format "owner/repo".}

\item{destdir}{Length-1 character vector; path to directory where to save the
output.}

\item{...}{Arguments passed to \code{\link[devtools:build]{devtools::build()}}.}
}
\value{
\itemize{
\item \code{download_github_zip()} outputs .zip files.
\item \code{build_github()} outputs source-package files (.tar.gz).
}
}
\description{
Build source package from source code in a GitHub repository.
}
\examples{
\dontrun{
urls <- c("maurolepore/toy1", "maurolepore/toy2")
tmp <- tempdir()

download_github_zip(urls, tmp)
fs::dir_ls(tmp, glob = "*.zip")

build_github(urls, tmp)
fs::dir_ls(tmp, glob = "*.tar.gz")
}
}
