% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_built_packages.R
\name{check_built_packages}
\alias{check_built_packages}
\title{Check multiple built source-packages.}
\usage{
check_built_packages(pkgs = install_src_pkgs())
}
\arguments{
\item{pkgs}{Character vector: The paths to built source-packages to
check.}
}
\value{
A \link[=tibble-package]{tibble}.
}
\description{
Check multiple built source-packages.
}
\examples{
\dontrun{
urls <- c("maurolepore/toy1", "maurolepore/toy2")
tmp <- tempdir()

build_github(urls, tmp)
pkgs <- fs::dir_ls(tmp, glob = "*.tar.gz")
pkgs

out <- check_built_packages()
out
}
}
