% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fgeo_source.R
\name{source_file}
\alias{source_file}
\title{Update vector to schedule package installation in correct order.}
\usage{
source_file(file, dir = "../fgeo.install")
}
\arguments{
\item{file}{Path to a file in data-raw/.}

\item{dir}{Path to the directory where \strong{fgeo.install} lives.}
}
\value{
Character vector.
}
\description{
Update vector to schedule package installation in correct order.
}
\examples{
\dontrun{
source_file("scheduled_packages")
source_file("fgeo_packages")
}
}
