% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fgeo_install.R
\name{update_fgeo_install}
\alias{update_fgeo_install}
\title{Update the key components (then run \code{check_built_packages()}).}
\usage{
update_fgeo_install()
}
\value{
Invisible \code{NULL}.
}
\description{
Update the key components (then run \code{check_built_packages()}).
}
\examples{
\dontrun{
update_fgeo_install()
out <- check_built_packages()
cat(out$value, sep = "\\n\\n")
}
}
