% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fgeo_source.R
\name{update_fgeo_source}
\alias{update_fgeo_source}
\title{Build fgeo source-packages from github into a local directory.}
\usage{
update_fgeo_source(pkgs = fgeo::fgeo(),
  src = fgeo.install::path_source())
}
\arguments{
\item{pkgs}{Character vector: The name of packages to update.}

\item{src}{Character vector: Path to directory where to save output.}
}
\value{
Source packages (.tar.gz files).
}
\description{
Build fgeo source-packages from github into a local directory.
}
\examples{
# Defaults
pkgs <- fgeo::fgeo()
src <- fgeo.install::path_source()
\dontrun{
update_fgeo_source(pkgs, src)
}
}
